requirements.txt
paramiko==4.0.0
paho-mqtt==2.1.0